Remove DatabaseConcertRepositoryBehaviors
It is not need anymore since ConcertRepository is a singleton
